refactor mapDictionaryInfo
