+++ packages/cspell-trie-lib/src/lib/mappers/mapDictionaryInfo.ts
@@ -1,12 +1,11 @@
 import { pipeSync } from '../../pipe';
-import { flatten, map, unique } from '../../pipe/operators';
-import type { DictionaryInformation, CharacterSetCosts } from '../models/DictionaryInformation';
+import { opFlatten, opMap } from '../../pipe/operators';
+import type { CharacterSetCosts, DictionaryInformation } from '../models/DictionaryInformation';
 import { parseLocale } from '../models/locale';
 import type { SuggestionCostMapDef } from '../models/suggestionCostsDef';
-import { accentForms, caseForms, expandCharacterSet } from '../utils/text';
-import { clean } from '../utils/util';
 import { EditCostsRequired, mapEditCosts } from './mapCosts';
-import { hunspellInformationToSuggestionCostDef, joinLetters } from './mapHunspellInformation';
+import { hunspellInformationToSuggestionCostDef } from './mapHunspellInformation';
+import { calcFirstLetterDef, parseAccents, parseAlphabet } from './mapToSuggestionCostDef';
 
 export function mapDictionaryInformation(dictInfo: DictionaryInformation): SuggestionCostMapDef[] {
     const _locale = dictInfo.locale;
@@ -36,51 +35,11 @@ function processAlphabet(
     return [
         ...pipeSync(
             csAlphabet,
-            map((cs) => parseAlphabet(cs, locale, editCost)),
-            flatten()
+            opMap((cs) => parseAlphabet(cs, locale, editCost)),
+            opFlatten()
         ),
         calcFirstLetterDef(csAlphabet, editCost),
     ];
-}
-
-function parseAlphabet(
-    cs: CharacterSetCosts,
-    locale: string[] | undefined,
-    editCost: EditCostsRequired
-): SuggestionCostMapDef[] {
-    const { cost, penalty } = cs;
-    const characters = expandCharacterSet(cs.characters);
-    const charForms = [
-        ...pipeSync(
-            characters,
-            map((c) => caseForms(c, locale))
-        ),
-    ];
-    const alphabet = joinLetters([
-        ...pipeSync(
-            charForms,
-            flatten(),
-            map((letter) => accentForms(letter)),
-            flatten(),
-            unique()
-        ),
-    ]);
-
-    const sugAlpha: SuggestionCostMapDef = clean({
-        map: alphabet,
-        replace: cost,
-        insDel: cost,
-        swap: cost,
-        penalty,
-    });
-
-    const caps = charForms.map((a) => joinLetters(a)).join('|');
-    const sugCaps: SuggestionCostMapDef = {
-        map: caps,
-        replace: editCost.capsCosts,
-    };
-
-    return [sugAlpha, sugCaps];
 }
 
 function toCharSets(
@@ -108,43 +67,10 @@ function toCharSets(
     return cs;
 }
 
-function calcFirstLetterDef(alphabet: CharacterSetCosts[], editCost: EditCostsRequired): SuggestionCostMapDef {
-    const mapOfFirstLetters = [
-        ...pipeSync(
-            alphabet,
-            map((cs) => cs.characters),
-            map((c) => expandCharacterSet(c)),
-            flatten(),
-            unique(),
-            map((letter) => `(^${letter})`)
-        ),
-    ].join('|');
-
-    const penalty = editCost.firstLetterPenalty;
-    // Make it a bit cheaper so it will match
-    const cost = editCost.baseCost - penalty;
-
-    return {
-        map: mapOfFirstLetters,
-        replace: cost,
-        penalty,
-    };
-}
-
 function processAccents(
     accents: DictionaryInformation['accents'],
     editCost: EditCostsRequired
 ): SuggestionCostMapDef[] {
     const cs = toCharSets(accents, '\u0300-\u0341', editCost.accentCosts);
     return cs.map((cs) => parseAccents(cs, editCost));
-}
-
-function parseAccents(cs: CharacterSetCosts, _editCost: EditCostsRequired): SuggestionCostMapDef {
-    const { cost, penalty } = cs;
-    const characters = expandCharacterSet(cs.characters);
-    return clean({
-        map: joinLetters([...characters]),
-        replace: cost,
-        penalty,
-    });
 }
